Reduce the packet size in UsbSend
The manual for the Silhouette Portrait indicate that some plotters will
produce skewed and distorted results (only really perceptible when you
try to use registration marks). One of the solutions in their manual is
to change the configuration in the Silhouette Studio to send 500 bytes at
a time, considering we are probably very far away from saturating the
USB communication while sending commands to the plotter, I take a far more
conservative approach of using the same size as the receiving block.